moving each example to its own page
