+++ config/services.yaml
@@ -10,8 +10,6 @@ services:
     _defaults:
         autowire: true      # Automatically injects dependencies in your services.
         autoconfigure: true # Automatically registers your services as commands, event subscribers, etc.
-        bind:
-            'Symfony\UX\TwigComponent\ComponentFactory $componentFactory': '@ux.twig_component.component_factory'
 
     # makes classes in src/ available to be used as services
     # this creates a service per class whose id is the fully-qualified class name
