+++ templates/main/example.html.twig
@@ -1,0 +1,49 @@
+{% extends 'base.html.twig' %}
+
+{% block title %}Live Twig Demo!{% endblock %}
+
+{% block body %}
+    <div>
+        <h1>
+            <code>{{ example.shortClass }}</code>
+            {% if example.isLive %}
+                <span
+                    class="badge bg-primary rounded-pill"
+                    style="font-size: 16px;"
+                    title="This component updates as your interact with it!"
+                >LIVE</span>
+            {% endif %}
+        </h1>
+        <div class="row">
+            <div class="col-6">
+                {{ include('examples/example_'~example.componentName~'.html.twig') }}
+            </div>
+            <div class="col-5">
+                <p>
+                    {{ example.classSummary|raw }}
+                </p>
+
+                <p>
+                    {{ example.classDescription|raw }}
+                </p>
+
+                <details>
+                    <summary><code>{{ example.shortClass }}.php</code></summary>
+
+                    <pre><code class="hljs php">{{ example_component_source(example.class) }}</code></pre>
+                </details>
+
+                <details>
+                    <summary><code>{{ example.componentName }}.html.twig</code></summary>
+
+                    <pre><code class="hljs twig">{{ example_component_template(example.componentName) }}</code></pre>
+                </details>
+
+                <div>
+                    Usage
+                    <pre><code class="hljs twig">{{ example_include_example_source(example.componentName) }}</code></pre>
+                </div>
+            </div>
+        </div>
+    </div>
+{% endblock %}
